iOS: Persist stacktraces of logged exceptions, improve ProgressViewController
