+++ common-jvm/src/main/java/org/ergoplatform/utils/LogUtils.kt
@@ -5,11 +5,13 @@ import java.util.*
 
 object LogUtils {
     var logDebug = false
+    var stackTraceLogger: ((String) -> Unit)? = null
 
     fun logDebug(tag: String, msg: String, t: Throwable? = null) {
         if (logDebug) {
             println(SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS").format(Date()) + " [" + tag + "] " + msg)
             t?.printStackTrace()
         }
+        t?.let { stackTraceLogger?.invoke(it.stackTraceToString()) }
     }
 }
